Move virtual Composite tests to a dedicated standalone class
The two SWT.VIRTUAL tests were added to Test_org_eclipse_swt_widgets_Composite,
which is a base class for ~70 widget test classes (Browser, CCombo, Table,
Tree, ...). As a result they were inherited and re-run for every subclass,
inflating suite runtime by ~5 minutes and creating a large amount of extra
widget churn that grew the GTK Display widget table to its 1024 threshold and
intermittently tripped a latent index-reuse error (g_object_get_qdata index -1)
in unrelated tests such as test_Constructor.

Move the tests into a standalone leaf class
(Test_org_eclipse_swt_widgets_Composite_Virtual) registered in AllWidgetTests
so they run exactly once, in the proper context. No production code change.
